Update dockerimagebuild.yml with modern language
We should all take a moment and consider why the word "master" has become so pervasive in our culture. Language matters. The words we choose matter.  More often than not there's a better word than one that implies a slave relationship. It's othering and unfair. Slavery is not okay, in any context or manipulation.

You don't have to change now, so we left "master" as the last branch name, but we prefer "prime" and many like "main".

Thank you for Listening. We Love You.